+++ packages/sdk/src/send/sdk-send.js
@@ -11,6 +11,7 @@ import {
     isGetBlockById,
     isGetBlockByHeight,
     isPing,
+    isGetCollection,
 } from "../interaction/interaction.js"
 import {sendTransaction} from "./send-transaction.js"
 import {sendGetTransactionStatus} from "./send-get-transaction-status.js"
@@ -23,6 +24,7 @@ import {sendGetBlockHeader} from "./send-get-block-header.js"
 import {sendGetLatestBlock} from "./send-get-latest-block.js"
 import {sendGetBlockById} from "./send-get-block-by-id.js"
 import {sendGetBlockByHeight} from "./send-get-block-by-height.js"
+import {sendGetCollection} from "./send-get-collection.js"
 import {sendPing} from "./send-ping.js"
 import {config} from "@onflow/config"
 
@@ -53,6 +55,8 @@ export const send = async (ix, opts = {}) => {
             return opts.sendGetBlockById ? opts.sendGetBlockById(ix, opts) : sendGetBlockById(ix, opts)
         case isGetBlockByHeight(ix):
             return opts.sendGetBlockByHeight ? opts.sendGetBlockByHeight(ix, opts) : sendGetBlockByHeight(ix, opts)
+        case isGetCollection(ix):
+            return opts.sendGetCollection ? opts.sendGetCollection(ix, opts) : sendGetCollection(ix, opts)
         case isPing(ix):
             return opts.sendPing ? opts.sendPing(ix, opts) : sendPing(ix, opts)
         default: